divide project into 3 components
- client
- server
- shared
